Traffic light problem complete!
Unfortunately, the folders would not commit, so I had to extract the individual files. If needed, here is the file hierarchy: Everything in a big folder ("traffic_light"), then a folder for the CSS ("css") and JavaScript ("js") respectively, both inside the original traffic_light folder. "style.css" was in "css", "script.js" was in "js", and finally, "index.html" was inside "traffic_light" but not in its own folder.